requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.3.0
minio>=7.2.0
anthropic>=0.40.0
httpx>=0.27.0
sentence-transformers>=3.0.0
pymupdf>=1.24.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
mcp>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
